+++ assets/javascripts/kitten/index.js
@@ -314,6 +314,7 @@ export { LinkBox } from './components/information/boxes/link-box'
 export { CommentAvatar } from './components/interaction/comment-block/comment-avatar'
 export { CommentForm } from './components/interaction/comment-block/comment-form'
 export { Comment } from './components/interaction/comment-block/comment'
+
 export { Discussion } from './components/interaction/discussion'
 export { DragAndDropList } from './components/interaction/drag-and-drop-list'
 export {
@@ -322,6 +323,7 @@ export {
   StepperIconInProgress,
   StepperIconValidated,
 } from './components/interaction/stepper-icon'
+export { EditorBar } from './components/interaction/editor-bar'
 export { Stepper } from './components/interaction/stepper'
 
 // Interaction/Upload